Reduce the api_log overhead in render passes
Some bevy users report that repetitively checking the global log level atomic adds up to a large overheadin wgpu  even when the log level is disabled (at runtime but still enable din the build). Since the vast majority of it tends to be in render passes with many commands, it's easy enough to check the global once per render pass and only skip the check for each individual command.